Address record refactoring.
Adding two new address records in place of the previous LocalgovAddress record:
- A generic UPRN-based address record and its interface.
- An Ordnance Survey Places address record and its interface.
